🎁 Refactor :roller to allow more options
Prior to this commit, I assumed a roller would be a function.

With this commit, I now allow a roller to be function, string, or
sequence.  We store that value in the roller then at "roll" time we
evaluate the roller via the `random-table/roller/roll' function.

This tightens the interface.